added comments and moved things to partial
