+++ Content.Client/Paper/UI/PaperBoundUserInterface.cs
@@ -1,24 +1,17 @@
 using JetBrains.Annotations;
-using Content.Client.Chat.TypingIndicator;
 using Robust.Client.UserInterface;
 using Robust.Client.UserInterface.Controls;
-using Content.Shared.Chat.TypingIndicator;
 using Robust.Shared.Utility;
 using Content.Shared.Paper;
 using static Content.Shared.Paper.PaperComponent;
-using Robust.Shared.Prototypes;
 
 namespace Content.Client.Paper.UI;
 
 [UsedImplicitly]
-public sealed class PaperBoundUserInterface : BoundUserInterface
+public sealed partial class PaperBoundUserInterface : BoundUserInterface
 {
     [ViewVariables]
     private PaperWindow? _window;
-
-    private static readonly ProtoId<TypingIndicatorPrototype> TypingIndicator = "paper";
-
-    private TypingIndicatorSystem? _typing;
 
     public PaperBoundUserInterface(EntityUid owner, Enum uiKey) : base(owner, uiKey)
     {
@@ -30,8 +23,8 @@ public sealed class PaperBoundUserInterface : BoundUserInterface
 
         _window = this.CreateWindow<PaperWindow>();
         _window.OnSaved += InputOnTextEntered;
-        _window.Typing += OnTyping;
-        _window.SubmitPressed += OnSubmit;
+        _window.Typing += OnTyping; // DeltaV
+        _window.SubmitPressed += OnSubmit; // DeltaV
 
         if (EntMan.TryGetComponent<PaperComponent>(Owner, out var paper))
         {
@@ -58,17 +51,5 @@ public sealed class PaperBoundUserInterface : BoundUserInterface
             _window.Input.TextRope = Rope.Leaf.Empty;
             _window.Input.CursorPosition = new TextEdit.CursorPos(0, TextEdit.LineBreakBias.Top);
         }
-    }
-
-    private void OnTyping()
-    {
-        _typing ??= EntMan.System<TypingIndicatorSystem>();
-        _typing?.ClientAlternateTyping(TypingIndicator);
-    }
-
-    private void OnSubmit()
-    {
-        _typing ??= EntMan.System<TypingIndicatorSystem>();
-        _typing?.ClientSubmittedChatText();
     }
 }
