+++ Content.Client/Paper/UI/PaperWindow.xaml.cs
@@ -49,8 +49,8 @@ namespace Content.Client.Paper.UI
         };
 
         public event Action<string>? OnSaved;
-        public event Action? Typing;
-        public event Action? SubmitPressed;
+        public event Action? Typing; // DeltaV
+        public event Action? SubmitPressed; // DeltaV
 
         private int _MaxInputLength = -1;
         public int MaxInputLength
@@ -77,20 +77,20 @@ namespace Content.Client.Paper.UI
             // Hook up the close button:
             CloseButton.OnPressed += _ =>
             {
-                SubmitPressed?.Invoke();
+                SubmitPressed?.Invoke(); // DeltaV
                 Close();
             };
 
             Input.OnKeyBindDown += args => // Solution while TextEdit don't have events
             {
-                Typing?.Invoke();
+                Typing?.Invoke(); // DeltaV
                 if (args.Function == EngineKeyFunctions.MultilineTextSubmit)
                 {
                     // SaveButton is disabled when we hit the max input limit. Just check
                     // that flag instead of trying to calculate the input length again
                     if (!SaveButton.Disabled)
                     {
-                        SubmitPressed?.Invoke();
+                        SubmitPressed?.Invoke(); // DeltaV
                         RunOnSaved();
                         args.Handle();
                     }
